requirements_streamlit.txt
streamlit>=1.24.0,<1.25.0
pandas>=1.5.3,<1.6.0
numpy>=1.22,<1.23
plotly>=5.15.0,<5.16.0
requests>=2.31.0,<2.32.0
matplotlib>=3.7.1,<3.8.0
seaborn>=0.12.2,<0.13.0
scikit-learn>=1.2.2,<1.3.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=1.10.9,<1.11.0